+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/life/BeanTransformerConfiguration.java
@@ -2,7 +2,6 @@ package com.dwarfeng.familyhelper.webapi.node.configuration.life;
 
 import com.dwarfeng.familyhelper.life.sdk.bean.entity.*;
 import com.dwarfeng.familyhelper.life.stack.bean.entity.*;
-import com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.life.disp.*;
 import com.dwarfeng.familyhelper.webapi.stack.bean.life.disp.*;
 import com.dwarfeng.subgrade.impl.bean.MapStructBeanTransformer;
@@ -15,59 +14,90 @@ public class BeanTransformerConfiguration {
 
     @Bean("life.pbSetBeanTransformer")
     public BeanTransformer<PbSet, JSFixedFastJsonPbSet> pbSetBeanTransformer() {
-        return new MapStructBeanTransformer<>(PbSet.class, JSFixedFastJsonPbSet.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                PbSet.class, JSFixedFastJsonPbSet.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPbSetBeanTransformer")
     public BeanTransformer<DispPbSet, JSFixedFastJsonDispPbSet> dispPbSetBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPbSet.class, JSFixedFastJsonDispPbSet.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPbSet.class, JSFixedFastJsonDispPbSet.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.popbBeanTransformer")
     public BeanTransformer<Popb, JSFixedFastJsonPopb> popbBeanTransformer() {
-        return new MapStructBeanTransformer<>(Popb.class, JSFixedFastJsonPopb.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Popb.class, JSFixedFastJsonPopb.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPopbBeanTransformer")
     public BeanTransformer<DispPopb, JSFixedFastJsonDispPopb> dispPopbBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPopb.class, JSFixedFastJsonDispPopb.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPopb.class, JSFixedFastJsonDispPopb.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.pbNodeBeanTransformer")
     public BeanTransformer<PbNode, JSFixedFastJsonPbNode> pbNodeBeanTransformer() {
-        return new MapStructBeanTransformer<>(PbNode.class, JSFixedFastJsonPbNode.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                PbNode.class, JSFixedFastJsonPbNode.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPbNodeBeanTransformer")
     public BeanTransformer<DispPbNode, JSFixedFastJsonDispPbNode> dispPbNodeBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPbNode.class, JSFixedFastJsonDispPbNode.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPbNode.class, JSFixedFastJsonDispPbNode.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.pbItemBeanTransformer")
     public BeanTransformer<PbItem, JSFixedFastJsonPbItem> pbItemBeanTransformer() {
-        return new MapStructBeanTransformer<>(PbItem.class, JSFixedFastJsonPbItem.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                PbItem.class, JSFixedFastJsonPbItem.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPbItemBeanTransformer")
     public BeanTransformer<DispPbItem, JSFixedFastJsonDispPbItem> dispPbItemBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPbItem.class, JSFixedFastJsonDispPbItem.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPbItem.class, JSFixedFastJsonDispPbItem.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.pbRecordBeanTransformer")
     public BeanTransformer<PbRecord, JSFixedFastJsonPbRecord> pbRecordBeanTransformer() {
-        return new MapStructBeanTransformer<>(PbRecord.class, JSFixedFastJsonPbRecord.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                PbRecord.class, JSFixedFastJsonPbRecord.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.pbFileInfoBeanTransformer")
     public BeanTransformer<PbFileInfo, JSFixedFastJsonPbFileInfo> pbFileInfoBeanTransformer() {
-        return new MapStructBeanTransformer<>(PbFileInfo.class, JSFixedFastJsonPbFileInfo.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                PbFileInfo.class, JSFixedFastJsonPbFileInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.activityTypeIndicatorBeanTransformer")
     public BeanTransformer<ActivityTypeIndicator, FastJsonActivityTypeIndicator>
     activityTypeIndicatorBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTypeIndicator.class, FastJsonActivityTypeIndicator.class, BeanMapper.class
+                ActivityTypeIndicator.class, FastJsonActivityTypeIndicator.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -75,7 +105,8 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ActivityTemplateDriverInfo, JSFixedFastJsonActivityTemplateDriverInfo>
     activityTemplateDriverInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplateDriverInfo.class, JSFixedFastJsonActivityTemplateDriverInfo.class, BeanMapper.class
+                ActivityTemplateDriverInfo.class, JSFixedFastJsonActivityTemplateDriverInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -83,14 +114,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ActivityTemplateDriverSupport, FastJsonActivityTemplateDriverSupport>
     activityTemplateDriverSupportBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplateDriverSupport.class, FastJsonActivityTemplateDriverSupport.class, BeanMapper.class
+                ActivityTemplateDriverSupport.class, FastJsonActivityTemplateDriverSupport.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("life.activityDataSetBeanTransformer")
     public BeanTransformer<ActivityDataSet, JSFixedFastJsonActivityDataSet> activityDataSetBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityDataSet.class, JSFixedFastJsonActivityDataSet.class, BeanMapper.class
+                ActivityDataSet.class, JSFixedFastJsonActivityDataSet.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -98,24 +131,32 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityDataSet, JSFixedFastJsonDispActivityDataSet>
     dispActivityDataSetBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityDataSet.class, JSFixedFastJsonDispActivityDataSet.class, BeanMapper.class
+                DispActivityDataSet.class, JSFixedFastJsonDispActivityDataSet.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.poadBeanTransformer")
     public BeanTransformer<Poad, JSFixedFastJsonPoad> poadBeanTransformer() {
-        return new MapStructBeanTransformer<>(Poad.class, JSFixedFastJsonPoad.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Poad.class, JSFixedFastJsonPoad.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPoadBeanTransformer")
     public BeanTransformer<DispPoad, JSFixedFastJsonDispPoad> dispPoadBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPoad.class, JSFixedFastJsonDispPoad.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPoad.class, JSFixedFastJsonDispPoad.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.activityDataNodeBeanTransformer")
     public BeanTransformer<ActivityDataNode, JSFixedFastJsonActivityDataNode> activityDataNodeBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityDataNode.class, JSFixedFastJsonActivityDataNode.class, BeanMapper.class
+                ActivityDataNode.class, JSFixedFastJsonActivityDataNode.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -123,14 +164,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityDataNode, JSFixedFastJsonDispActivityDataNode>
     dispActivityDataNodeBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityDataNode.class, JSFixedFastJsonDispActivityDataNode.class, BeanMapper.class
+                DispActivityDataNode.class, JSFixedFastJsonDispActivityDataNode.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.activityDataItemBeanTransformer")
     public BeanTransformer<ActivityDataItem, JSFixedFastJsonActivityDataItem> activityDataItemBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityDataItem.class, JSFixedFastJsonActivityDataItem.class, BeanMapper.class
+                ActivityDataItem.class, JSFixedFastJsonActivityDataItem.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -138,14 +181,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityDataItem, JSFixedFastJsonDispActivityDataItem>
     dispActivityDataItemBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityDataItem.class, JSFixedFastJsonDispActivityDataItem.class, BeanMapper.class
+                DispActivityDataItem.class, JSFixedFastJsonDispActivityDataItem.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.activityDataRecordBeanTransformer")
     public BeanTransformer<ActivityDataRecord, JSFixedFastJsonActivityDataRecord> activityDataRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityDataRecord.class, JSFixedFastJsonActivityDataRecord.class, BeanMapper.class
+                ActivityDataRecord.class, JSFixedFastJsonActivityDataRecord.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -153,14 +198,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityDataRecord, JSFixedFastJsonDispActivityDataRecord>
     dispActivityDataRecordBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityDataRecord.class, JSFixedFastJsonDispActivityDataRecord.class, BeanMapper.class
+                DispActivityDataRecord.class, JSFixedFastJsonDispActivityDataRecord.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.activityTemplateBeanTransformer")
     public BeanTransformer<ActivityTemplate, JSFixedFastJsonActivityTemplate> activityTemplateBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplate.class, JSFixedFastJsonActivityTemplate.class, BeanMapper.class
+                ActivityTemplate.class, JSFixedFastJsonActivityTemplate.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -168,35 +215,49 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityTemplate, JSFixedFastJsonDispActivityTemplate>
     dispActivityTemplateBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityTemplate.class, JSFixedFastJsonDispActivityTemplate.class, BeanMapper.class
+                DispActivityTemplate.class, JSFixedFastJsonDispActivityTemplate.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.poatBeanTransformer")
     public BeanTransformer<Poat, JSFixedFastJsonPoat> poatBeanTransformer() {
-        return new MapStructBeanTransformer<>(Poat.class, JSFixedFastJsonPoat.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Poat.class, JSFixedFastJsonPoat.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPoatBeanTransformer")
     public BeanTransformer<DispPoat, JSFixedFastJsonDispPoat> dispPoatBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPoat.class, JSFixedFastJsonDispPoat.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPoat.class, JSFixedFastJsonDispPoat.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.poatacBeanTransformer")
     public BeanTransformer<Poatac, JSFixedFastJsonPoatac> poatacBeanTransformer() {
-        return new MapStructBeanTransformer<>(Poatac.class, JSFixedFastJsonPoatac.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Poatac.class, JSFixedFastJsonPoatac.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPoatacBeanTransformer")
     public BeanTransformer<DispPoatac, JSFixedFastJsonDispPoatac> dispPoatacBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPoatac.class, JSFixedFastJsonDispPoatac.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPoatac.class, JSFixedFastJsonDispPoatac.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.activityTemplateCoverInfoBeanTransformer")
     public BeanTransformer<ActivityTemplateCoverInfo, JSFixedFastJsonActivityTemplateCoverInfo>
     activityTemplateCoverInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplateCoverInfo.class, JSFixedFastJsonActivityTemplateCoverInfo.class, BeanMapper.class
+                ActivityTemplateCoverInfo.class, JSFixedFastJsonActivityTemplateCoverInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -204,7 +265,8 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ActivityTemplateDataInfo, JSFixedFastJsonActivityTemplateDataInfo>
     activityTemplateDataInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplateDataInfo.class, JSFixedFastJsonActivityTemplateDataInfo.class, BeanMapper.class
+                ActivityTemplateDataInfo.class, JSFixedFastJsonActivityTemplateDataInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -214,7 +276,7 @@ public class BeanTransformerConfiguration {
         return new MapStructBeanTransformer<>(
                 DispActivityTemplateDataInfo.class,
                 JSFixedFastJsonDispActivityTemplateDataInfo.class,
-                BeanMapper.class
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
@@ -224,7 +286,7 @@ public class BeanTransformerConfiguration {
         return new MapStructBeanTransformer<>(
                 ActivityTemplateParticipant.class,
                 JSFixedFastJsonActivityTemplateParticipant.class,
-                BeanMapper.class
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -234,7 +296,7 @@ public class BeanTransformerConfiguration {
         return new MapStructBeanTransformer<>(
                 DispActivityTemplateParticipant.class,
                 JSFixedFastJsonDispActivityTemplateParticipant.class,
-                BeanMapper.class
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
@@ -242,38 +304,50 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ActivityTemplateFileInfo, JSFixedFastJsonActivityTemplateFileInfo>
     activityTemplateFileInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityTemplateFileInfo.class, JSFixedFastJsonActivityTemplateFileInfo.class, BeanMapper.class
+                ActivityTemplateFileInfo.class, JSFixedFastJsonActivityTemplateFileInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
     @Bean("life.activityBeanTransformer")
     public BeanTransformer<Activity, JSFixedFastJsonActivity> activityBeanTransformer() {
-        return new MapStructBeanTransformer<>(Activity.class, JSFixedFastJsonActivity.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Activity.class, JSFixedFastJsonActivity.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispActivityBeanTransformer")
     public BeanTransformer<DispActivity, JSFixedFastJsonDispActivity>
     dispActivityBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivity.class, JSFixedFastJsonDispActivity.class, BeanMapper.class
+                DispActivity.class, JSFixedFastJsonDispActivity.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.poacBeanTransformer")
     public BeanTransformer<Poac, JSFixedFastJsonPoac> poacBeanTransformer() {
-        return new MapStructBeanTransformer<>(Poac.class, JSFixedFastJsonPoac.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                Poac.class, JSFixedFastJsonPoac.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
+        );
     }
 
     @Bean("life.dispPoacBeanTransformer")
     public BeanTransformer<DispPoac, JSFixedFastJsonDispPoac> dispPoacBeanTransformer() {
-        return new MapStructBeanTransformer<>(DispPoac.class, JSFixedFastJsonDispPoac.class, BeanMapper.class);
+        return new MapStructBeanTransformer<>(
+                DispPoac.class, JSFixedFastJsonDispPoac.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
+        );
     }
 
     @Bean("life.activityCoverInfoBeanTransformer")
     public BeanTransformer<ActivityCoverInfo, JSFixedFastJsonActivityCoverInfo>
     activityCoverInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityCoverInfo.class, JSFixedFastJsonActivityCoverInfo.class, BeanMapper.class
+                ActivityCoverInfo.class, JSFixedFastJsonActivityCoverInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -281,7 +355,8 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ActivityParticipant, JSFixedFastJsonActivityParticipant>
     activityParticipantBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityParticipant.class, JSFixedFastJsonActivityParticipant.class, BeanMapper.class
+                ActivityParticipant.class, JSFixedFastJsonActivityParticipant.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 
@@ -289,14 +364,16 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<DispActivityParticipant, JSFixedFastJsonDispActivityParticipant>
     dispActivityParticipantBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                DispActivityParticipant.class, JSFixedFastJsonDispActivityParticipant.class, BeanMapper.class
+                DispActivityParticipant.class, JSFixedFastJsonDispActivityParticipant.class,
+                com.dwarfeng.familyhelper.webapi.sdk.bean.life.BeanMapper.class
         );
     }
 
     @Bean("life.activityFileInfoBeanTransformer")
     public BeanTransformer<ActivityFileInfo, JSFixedFastJsonActivityFileInfo> activityFileInfoBeanTransformer() {
         return new MapStructBeanTransformer<>(
-                ActivityFileInfo.class, JSFixedFastJsonActivityFileInfo.class, BeanMapper.class
+                ActivityFileInfo.class, JSFixedFastJsonActivityFileInfo.class,
+                com.dwarfeng.familyhelper.life.sdk.bean.BeanMapper.class
         );
     }
 }